"Fix" mvm_rottenburg_rev_int_spanner_switchup.pop
w4:
removed interruptaction because it is terrible and broken and terrible and broken and terrible. (i have spent hours trying to fix it ok sorry)
DF pyros moved to after the tank dies.
tank hp nerfed from 15k to 10k to compensate for the increased dps requirement over the intended (and judged) mission.